fixes bug on SGE reported by Alex
I had an issue the other day with standalone on the SGE cluster. It was
picking up SGE_TASK_ID to be "undefined", so asking me to set the number
of processes in the group. If I did this (as 1), it would break saying
that the job_idx isn't an integer.